Map and Player input changes
-Made level more difficult in later areas and added more collectibles to the map
-Player input is now hold, rather than repeated tapping
-Fixed issue with piss meter not being clamped

+++ Assets/Characters/Buster/BusterController.cs
@@ -6,7 +6,7 @@ public class BusterController : MonoBehaviour
 {
     public Rigidbody rb;
 
-    public int speeeeed = 1;
+    public float speeeeed = 1;
     public Vector2 inputDelay;
 
     public bool canMove = true;
